Move the helpers back into pear
This makes it easier to ship this in core later
